lib/serverpool: allow 32-bit systems to pick a random server
There's no need to specify a uint64_t rather than size_t here.

+++ lib/serverpool/serverpool.c
@@ -430,7 +430,7 @@ err0:
 struct sock_addr *
 serverpool_pick(struct serverpool * P)
 {
-	uint64_t i;
+	size_t i;
 
 	/* Get rid of any expired addresses. */
 	if (pruneaddrs(P))
